[🕹️refactor]  :  타입 컴포넌트 분기로 인한 상대주소 수정
